# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: C5 360相机巡查标注系统 API
  version: 1.0.0
  license:
    name: Proprietary
  description: >
    Spec-first OpenAPI 3 contract for project C5 (360-camera inspection &
    annotation system). This single document is the authoritative source for
    oapi-codegen v2 → Go server stubs + TypeScript + Kotlin/Retrofit clients.


    Conventions:

    - Every response uses the shared envelope `{success, data, error, meta?}`.
      List endpoints populate `meta` with `{total, page, page_size}`.

    - All IDs are int64 (BIGSERIAL in the DB). `client_uuid` / `job_uuid` are
      UUID strings (offline idempotency / public job handles).

    - All geometry on the wire is WGS84 / SRID 4326 GeoJSON. GCJ-02 is NEVER
      persisted or transmitted; it is a client-only render-boundary concern.

    - All timestamps are RFC3339 / ISO-8601 UTC (`format: date-time`).

    - Error codes are a frozen, canonical enum (see ErrorObject.code).


    Out-of-band (NOT expressible in OpenAPI):

    - `GET /exports/{job_uuid}/events` is a Server-Sent Events (`text/event-stream`)
      endpoint emitting `event: progress` and a terminal `event: done`. It is
      hand-written with `GET /exports/{job_uuid}` as the polling fallback
      (Tencent CLB/CDN may buffer SSE). It is intentionally absent from `paths`.

    - Dictionary / config pull uses ETag (`If-None-Match: "<content_hash>"` →
      304 Not Modified when unchanged). 304 responses carry no body.

...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /media/upload-credentials:
    post:
      tags: [media]
      operationId: createMediaUploadCredentials
      summary: Issue STS credentials scoped to a per-upload prefix (6 multipart actions)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaUploadCredentialsRequest'
      responses:
        '200':
          description: STS credentials + COS target
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaUploadCredentialsResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /media/confirm:
    post:
      tags: [media]
      operationId: confirmMedia
      summary: HeadObject-verify an uploaded object, advance to CONFIRMED + set verified_at
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaConfirmRequest'
      responses:
        '200':
          description: Media confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAssetResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/MediaVerifyFailed'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '500':
          $ref: '#/components/responses/InternalError'
  /media/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [media]
      operationId: getMedia
      summary: Get a media asset with a signed CDN URL for its tier
      responses:
        '200':
          description: Media asset (signed url)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAssetResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Stats
  # ===========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access JWT (golang-jwt/v5). Refresh is an opaque Redis token.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
...
    Unauthenticated:
      description: Missing/invalid/expired bearer token (UNAUTHENTICATED | TOKEN_EXPIRED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: casbin (sub,obj,act) denied (FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or soft-deleted (NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    MediaVerifyFailed:
      description: HeadObject mismatch (MEDIA_VERIFY_FAILED); state stays UPLOADING
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: Boundary schema / business validation failed (VALIDATION_FAILED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Hot-path rate limit hit (RATE_LIMITED); client must backoff
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unhandled error (INTERNAL); never leaks stack/SQL
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Meta:
      type: object
      properties:
        total: { type: integer, format: int64 }
        page: { type: integer }
        page_size: { type: integer }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - UNAUTHENTICATED
        - TOKEN_EXPIRED
        - FORBIDDEN
        - NOT_FOUND
        - CONFLICT
        - MEDIA_VERIFY_FAILED
        - DICT_VERSION_RETIRED
        - RATE_LIMITED
        - INTERNAL
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message: { type: string }
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: false }
        data:
          type: object
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    CaptureState:
      type: string
      enum: [CAPTURED_RAW, STITCHED, QUEUED, UPLOADING, UPLOADED, CONFIRMED]
    MediaTier:
      type: string
      enum: [original, web, thumb]
    MediaOwnerType:
      type: string
      enum: [problem, inspection, project, user]
...
    MediaAsset:
      type: object
      required:
        - id
        - client_uuid
        - owner_type
        - owner_id
        - tier
        - cos_bucket
        - cos_key
        - cos_region
        - capture_state
        - created_at
        - updated_at
      properties:
        id: { type: integer, format: int64 }
        client_uuid: { type: string, format: uuid }
        owner_type:
          $ref: '#/components/schemas/MediaOwnerType'
        owner_id: { type: integer, format: int64 }
        tier:
          $ref: '#/components/schemas/MediaTier'
        cos_bucket: { type: string }
        cos_key: { type: string }
        cos_region: { type: string }
        content_type: { type: string }
        byte_size: { type: integer, format: int64, nullable: true }
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        etag: { type: string, nullable: true }
        capture_state:
          $ref: '#/components/schemas/CaptureState'
        verified_at: { type: string, format: date-time, nullable: true }
        media_group: { type: string, format: uuid, nullable: true }
        meta:
          type: object
          additionalProperties: true
        signed_url:
          type: string
          nullable: true
          description: Signed CDN URL for this tier (populated by GET /media/{id}).
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    MediaUploadCredentialsRequest:
      type: object
      required: [owner_type, owner_id, tier, content_type, byte_size, client_uuid]
      properties:
        owner_type:
          $ref: '#/components/schemas/MediaOwnerType'
        owner_id: { type: integer, format: int64 }
        tier:
          $ref: '#/components/schemas/MediaTier'
        content_type: { type: string }
        byte_size: { type: integer, format: int64 }
        client_uuid: { type: string, format: uuid }
    MediaConfirmRequest:
      type: object
      required: [client_uuid, key, etag, byte_size]
      properties:
        client_uuid: { type: string, format: uuid }
        key:
          type: string
          description: COS object key under the per-upload prefix.
        etag: { type: string }
        byte_size: { type: integer, format: int64 }
        width: { type: integer }
        height: { type: integer }
...
    MediaUploadCredentials:
      type: object
      required: [bucket, region, key, prefix, credentials]
      properties:
        bucket: { type: string }
        region: { type: string }
        key: { type: string }
        prefix:
          type: string
          description: STS policy is scoped to this per-upload prefix.
        credentials:
          $ref: '#/components/schemas/StsCredentials'
    StsCredentials:
      type: object
      required: [tmpSecretId, tmpSecretKey, sessionToken, expiredTime]
      properties:
        tmpSecretId: { type: string }
        tmpSecretKey: { type: string }
        sessionToken: { type: string }
        expiredTime: { type: integer, format: int64 }

    # -------------------------------------------------------------------------
    # Per-endpoint typed response wrappers
    # -------------------------------------------------------------------------
...
    MediaAssetResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/MediaAsset'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
    MediaUploadCredentialsResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/MediaUploadCredentials'
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...